refactor: chang id type

+++ common/src/main/java/com/example/addictions/dto/AddictionDto.java
@@ -1,10 +1,9 @@
 package com.example.addictions.dto;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 public class AddictionDto {
-    private int id;
+    private long id;
     private String name;
     private float costPerDay;
     private String deadline;
